Formatting with ruff
