+++ examples/prof.py
@@ -288,7 +288,7 @@ def prof_dense_qp(trial, nz, nbatch, cons, cuda=True):
         'nz': nz,
         'nbatch': nbatch,
         'cuda': cuda,
-        'mode': f'cvxpylayers',
+        'mode': 'cvxpylayers',
         'direction': 'forward',
         'time': t,
         'qp': 'dense',
@@ -304,7 +304,7 @@ def prof_dense_qp(trial, nz, nbatch, cons, cuda=True):
         'nz': nz,
         'nbatch': nbatch,
         'cuda': cuda,
-        'mode': f'cvxpylayers',
+        'mode': 'cvxpylayers',
         'direction': 'backward',
         'time': t,
         'qp': 'dense',
