+++ src/cvxpylayers/jax/cvxpylayer.py
@@ -1,9 +1,7 @@
-import diffcp
 import cvxpy as cp
 from cvxpy.reductions.solvers.conic_solvers.scs_conif import \
     dims_to_solver_dict
 import numpy as np
-import time
 from functools import partial
 from cvxpylayers.utils import \
     ForwardContext, BackwardContext, forward_numpy, backward_numpy
